Fix incorrect local timestamp parsing
Some Reddit endpoints use attributes with both a local and a UTC
timestamp, for example `created` and `created_utc` in a post. The local
timestamp supposedly is only kept for legacy reasons and should not be
used because it adds an unpredictable timezone offset. The library
incorrectly parses it as a UTC timestamp, and then uses it as the main
value to build the `Controllers` objects, leading to the controller
properties representing the wrong moment in time.

This fixes the issue by both parsing the local timestamps correctly
assuming the host's local timezone offset and by using the UTC values to
build `Controller` objects to avoid unpredictable behavior. Local
timestamps in `Things` objects are also marked as obsolete in case
anyone was using them directly.

Fixes #68